+++ services/121-service/src/notifications/twilio.client.ts
@@ -1,4 +1,10 @@
 import { TWILIO } from '../secrets';
 
-export const twilioClient = require('twilio')(TWILIO.sid, TWILIO.authToken);
+export const twilioClient = require('twilio')(
+  TWILIO.tokenSid,
+  TWILIO.tokenSecret,
+  {
+    accountSid: TWILIO.sid,
+  },
+);
 export const twilio = require('twilio');